+++ packages/widgetbook/lib/src/fields/fields.dart
@@ -5,5 +5,6 @@ export 'double_slider_field.dart';
 export 'field.dart';
 export 'field_codec.dart';
 export 'field_type.dart';
+export 'fields_composable.dart';
 export 'list_field.dart';
 export 'string_field.dart';